+++ Sources/BraintreeThreeDSecure/BTPaymentFlowClient+ThreeDSecure.m
@@ -1,8 +1,5 @@
 #import "BTPaymentFlowClient+ThreeDSecure_Internal.h"
-#import "BTThreeDSecureResult_Internal.h"
 #import "BTThreeDSecureRequest_Internal.h"
-#import "BTThreeDSecurePostalAddress_Internal.h"
-#import "BTThreeDSecureAdditionalInformation_Internal.h"
 
 // MARK: - Objective-C File Imports for Package Managers
 #if __has_include(<Braintree/BraintreeThreeDSecure.h>) // CocoaPods
@@ -40,6 +37,13 @@
 
 #else                                            // Carthage
 #import <BraintreeCore/BraintreeCore-Swift.h>
+#endif
+
+// MARK: - Temporary Swift Module Imports
+#if __has_include(<Braintree/BraintreeThreeDSecure.h>) // CocoaPods
+#import <Braintree/Braintree-Swift.h>
+#else                                            // SPM and Carthage
+#import <BraintreeThreeDSecure/BraintreeThreeDSecure-Swift.h>
 #endif
 
 @implementation BTPaymentFlowClient (ThreeDSecure)
@@ -153,7 +157,16 @@ NSString * const BTThreeDSecureFlowValidationErrorsKey = @"com.braintreepayments
                 return;
             }
 
-            completionBlock([[BTThreeDSecureResult alloc] initWithJSON:body], nil);
+            if (body) {
+                completionBlock([[BTThreeDSecureResult alloc] initWithJSON:body], nil);
+                return;
+            } else {
+                error = [NSError errorWithDomain:BTThreeDSecureFlowErrorDomain
+                                            code:BTThreeDSecureFlowErrorTypeNoBodyReturned
+                                        userInfo:@{NSLocalizedDescriptionKey: @"A body was not returned from the API during the request."}];
+                completionBlock(nil, error);
+                return;
+            }
         }];
     }];
 }