+++ UnitTests/BraintreeThreeDSecureTests/BraintreeThreeDSecureTests-Bridging-Header.h
@@ -1,9 +1,4 @@
-#import <BraintreeThreeDSecure/BTThreeDSecureResult_Internal.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureRequest_Internal.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureLookup_Internal.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureAuthenticateJWT.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureAdditionalInformation_Internal.h>
-#import <BraintreeThreeDSecure/BTThreeDSecurePostalAddress_Internal.h>
 #import <BraintreeThreeDSecure/BTPaymentFlowClient+ThreeDSecure_Internal.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureRequest.h>
 #import <BraintreePaymentFlow/BTPaymentFlowClient.h>